Add Get Session for Listener, fix sg logging, pulls to results dir now, add test stub
