+++ resources/sync_gateway_configs/walrus-revs-limit.json
@@ -1,0 +1,14 @@
+{
+	"interface":":4984",
+    "adminInterface": "0.0.0.0:4985",
+	"log": ["*"],
+	"databases": {
+		"db": {
+			"revs_limit" : 10,
+			"server": "walrus:",
+			"sync":`function(doc){
+				channel(doc.channels);
+			}`,
+		}
+	}
+}
