+++ testsuites/syncgateway/functional/1sg_1cbs/1sg_1cbs-Part1.robot
@@ -2,7 +2,7 @@
 Resource    resources/common.robot
 
 Library     ${Libraries}/NetworkUtils.py
-Library     ${Libraries}/LoggingKeywords.py
+Library     ${KEYWORDS}/Logging.py
 
 Library     ../test_bucket_shadow.py
 Library     ../test_bulk_get_compression.py
