+++ testsuites/syncgateway/functional/1sg_2ac_1cbs/1sg_2ac_1cbs.robot
@@ -4,7 +4,7 @@ Resource    resources/common.robot
 Library     Process
 Library     OperatingSystem
 Library     ${Libraries}/NetworkUtils.py
-Library     ${Libraries}/LoggingKeywords.py
+Library     ${KEYWORDS}/Logging.py
 
 Library     ../test_cbgt_pindex.py
 Library     ../test_dcp_reshard.py
